backend/pyproject.toml
[project]
name = "dm-toolset-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-ninja>=1.4.3",
    "pillow>=11.3.0",
    "psycopg[binary,pool]>=3.2.9",
    "python-dotenv>=1.1.1",
]
